+++ GFT-Podcasts/Controllers/PodcastController.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Linq;
 using GFT_Podcasts.Libraries.ExtensionsMethods;
 using GFT_Podcasts.Models;
 using GFT_Podcasts.Models.ViewModels;
@@ -7,14 +6,19 @@ using GFT_Podcasts.Models.ViewModels.PodcastViewModels;
 using GFT_Podcasts.Repositories.Interfaces;
 using Microsoft.AspNetCore.Mvc;
 
-namespace GFT_Podcasts.Controllers {
+namespace GFT_Podcasts.Controllers
+{
     [Route("api/")]
     public class PodcastController : ControllerBase {
         private readonly IPodcastRepository _podcastRepository;
+
+        private readonly ICategoriaRepository _categoriaRepository;
         
-        public PodcastController(IPodcastRepository podcastRepository) {
+        public PodcastController(IPodcastRepository podcastRepository, ICategoriaRepository categoriaRepository) {
             _podcastRepository = podcastRepository;
+            _categoriaRepository= categoriaRepository;
         }
+        
 
         [HttpGet]
         [Route("v1/podcasts/{id}")]
@@ -41,15 +45,23 @@ namespace GFT_Podcasts.Controllers {
         [Route("v1/podcasts")]
         public ObjectResult Get() {
             var podcasts = _podcastRepository.Listar();
-            return new ObjectResult(new ResultViewModel(true, "Listagem de Podcasts!", podcasts));
+            return !podcasts.Any()
+                ? new ObjectResult
+                    (new ResultViewModel(false, "Nenhum podcast encontrado.", podcasts))
+                : new ObjectResult
+                    (new ResultViewModel(true, "Listagem de Podcasts!", podcasts));
         }
         
         [HttpPost]
         [Route("v1/podcasts/")]
         public ObjectResult Post([FromBody] PodcastCadastroViewModel podcastTemp) {
-            if (!ModelState.IsValid)
+            if (!_categoriaRepository.Existe(podcastTemp.CategoriaId)) {
+                ModelState.AddModelError("CategoriaId", "Categoria inexistente.");
+            }
+            if (!ModelState.IsValid) {
                 return new ObjectResult(new ResultViewModel(false, "Erro ao cadastrar podcast.",
                     ModelState.ListarErros()));
+            }
             var podcast = new Podcast() {
                 Id = 0,
                 Nome = podcastTemp.Nome,
@@ -61,7 +73,7 @@ namespace GFT_Podcasts.Controllers {
             };
             _podcastRepository.Criar(podcast);
             return new ObjectResult(new ResultViewModel(true, "Podcast cadastrado com sucesso!", 
-                podcast));
+                podcastTemp));
         }
         
         [HttpPut]
@@ -70,7 +82,12 @@ namespace GFT_Podcasts.Controllers {
             if (id != podcastTemp.Id) {
                 ModelState.AddModelError("Id", "Id da requisição difere do Id da categoria.");
             }
-            
+            if (!_podcastRepository.Existe(podcastTemp.Id)) {
+                ModelState.AddModelError("Id", "Podcast inexistente.");
+            }
+            if (!_categoriaRepository.Existe(podcastTemp.CategoriaId)) {
+                ModelState.AddModelError("CategoriaId", "Categoria inexistente.");
+            }
             if (!ModelState.IsValid)
                 return new ObjectResult(new ResultViewModel(false, "Erro ao editar categoria.",
                     ModelState.ListarErros()));
@@ -85,16 +102,20 @@ namespace GFT_Podcasts.Controllers {
             };
             _podcastRepository.Editar(podcast);
             return new ObjectResult(new ResultViewModel(true, "Podcast editado com sucesso!", 
-                podcast));
+                podcastTemp));
         }
         
         [HttpDelete]
         [Route("v1/podcasts/{id}")]
         public ObjectResult Delete(int id) {
             var podcast = _podcastRepository.Buscar(id);
+            if (podcast == null) {
+                return new ObjectResult(new ResultViewModel(false, "Podcast inexistente.", null));
+            }
             _podcastRepository.Remover(podcast);
             return new ObjectResult(new ResultViewModel(true, "Podcast excluído com sucesso!", 
                 podcast));
+
         }
     }
 }
